+++ src/main/kotlin/hr/eduwalk/data/model/LocationScore.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class LocationScore(
     val userId: String,
-    val locationId: Int,
+    val locationId: Long,
     val score: Int? = null,
 )
