+++ src/main/kotlin/hr/eduwalk/domain/interfaces/IQuestionDao.kt
@@ -6,6 +6,6 @@ import hr.eduwalk.domain.model.ServiceResult
 interface IQuestionDao {
     suspend fun insertQuestion(question: Question): ServiceResult<Unit>
     suspend fun updateQuestion(question: Question): ServiceResult<Boolean>
-    suspend fun deleteQuestion(questionId: Int): ServiceResult<Unit>
-    suspend fun getQuestionsForLocation(locationId: Int): ServiceResult<List<Question>>
+    suspend fun deleteQuestion(questionId: Long): ServiceResult<Unit>
+    suspend fun getQuestionsForLocation(locationId: Long): ServiceResult<List<Question>>
 }
