+++ src/main/kotlin/hr/eduwalk/domain/usecase/locationscore/GetLocationScoreForUser.kt
@@ -7,6 +7,6 @@ import hr.eduwalk.domain.model.toLocationScoreResponse
 class GetLocationScoreForUser(
     private val locationScoreDao: ILocationScoreDao,
 ) {
-    suspend operator fun invoke(locationId: Int, userId: String): LocationScoreResponse =
+    suspend operator fun invoke(locationId: Long, userId: String): LocationScoreResponse =
         locationScoreDao.getLocationScoreForUser(locationId = locationId, userId = userId).toLocationScoreResponse()
 }
