+++ src/main/kotlin/hr/eduwalk/domain/usecase/question/DeleteQuestion.kt
@@ -7,6 +7,6 @@ import hr.eduwalk.domain.model.toEmptyResponse
 class DeleteQuestion(
     private val questionDao: IQuestionDao,
 ) {
-    suspend operator fun invoke(questionId: Int): EmptyResponse =
+    suspend operator fun invoke(questionId: Long): EmptyResponse =
         questionDao.deleteQuestion(questionId = questionId).toEmptyResponse()
 }
